Fixed rpc tests after disabling concent
