+++ scripts/node_integration_tests/playbooks/golem/rpc_test/concent/test_config.py
@@ -1,0 +1,8 @@
+from ....test_config_base import TestConfigBase
+
+
+class TestConfig(TestConfigBase):
+    def __init__(self):
+        super().__init__()
+        for node_config in self.nodes.values():
+            node_config.concent = 'staging'
