+++ scripts/node_integration_tests/playbooks/golem/rpc_test/playbook.py
@@ -6,12 +6,8 @@ from ...test_config_base import NodeId
 
 class Playbook(NodeTestPlaybook):
     def step_verify_deposit_balance_call(self):
-        def on_success(result):
-            if result['value'] and result['status'] and result['timelock']:
-                print("Result correct %s" % result)
-                self.next()
-            else:
-                print("Unexpected result: %s" % result)
+        def on_success(_):
+            self.next()
 
         def on_error(error):
             self.fail(error)
