--- scripts/node_integration_tests/playbooks/golem/rpc_test/testnet/test_config.py
@@ -1,5 +1,0 @@
-from ....test_config_base import TestConfigBase
-
-
-class TestConfig(TestConfigBase):
-    pass
